refactor: remove utils crates
