+++ cli/src/tests/deployment.rs
@@ -1,5 +1,6 @@
 use crate::deploy::eigenlayer::{EigenlayerDeployOpts, deploy_avs_contracts};
 use alloy_provider::RootProvider;
+use blueprint_evm_extra::util::get_provider_http;
 use blueprint_runner::config::SupportedChains;
 use color_eyre::eyre::Result;
 use gadget_logging::setup_log;
@@ -7,7 +8,6 @@ use gadget_std::collections::HashMap;
 use gadget_std::fs;
 use gadget_std::process::Command;
 use gadget_testing_utils::anvil::start_default_anvil_testnet;
-use gadget_utils::evm::get_provider_http;
 use serde_json::Value;
 use tempfile::TempDir;
