+++ crates/clients/evm/Cargo.toml
@@ -41,7 +41,7 @@ gadget-logging = { workspace = true }
 gadget-rpc-calls = { workspace = true }
 
 [dev-dependencies]
-gadget-utils-evm = { workspace = true }
+blueprint-evm-extra = { workspace = true }
 gadget-anvil-testing-utils = { workspace = true }
 alloy-signer-local = { workspace = true }
 testcontainers = { workspace = true }
