+++ crates/eigenlayer-extra/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "gadget-utils-evm"
+name = "blueprint-eigenlayer-extra"
 version = "0.1.0"
-description = "EVM utilities for Tangle Blueprints"
+description = "EigenLayer utilites for Tangle Blueprints"
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -12,18 +12,18 @@ repository.workspace = true
 workspace = true
 
 [dependencies]
-gadget-std = { workspace = true }
+blueprint-evm-extra = { workspace = true }
+alloy-contract = { workspace = true, default-features = false }
 alloy-network = { workspace = true, default-features = false }
 alloy-primitives = { workspace = true, default-features = false }
 alloy-provider = { workspace = true, default-features = false }
-alloy-signer = { workspace = true, default-features = false }
-alloy-signer-local = { workspace = true, default-features = false }
 alloy-transport = { workspace = true, default-features = false }
-url = { workspace = true }
+eigensdk = { workspace = true, default-features = false, features = ["utils"] }
+thiserror = { workspace = true }
 
 [features]
 default = ["std"]
 std = [
+    "blueprint-evm-extra/std",
     "alloy-primitives/std",
-    "gadget-std/std",
 ]
