+++ crates/eigenlayer-extra/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod util;
