+++ crates/evm-extra/src/lib.rs
@@ -10,7 +10,6 @@
 #![doc(
     html_logo_url = "https://cdn.prod.website-files.com/6494562b44a28080aafcbad4/65aaf8b0818b1d504cbdf81b_Tnt%20Logo.png"
 )]
-#![warn(missing_docs)]
 #![deny(
     missing_debug_implementations,
     missing_copy_implementations,
@@ -19,8 +18,7 @@
     unsafe_code,
     unstable_features,
     unused_import_braces,
-    unused_qualifications,
-    missing_docs
+    unused_qualifications
 )]
 
 extern crate alloc;
@@ -29,3 +27,4 @@ pub mod consumer;
 pub mod extract;
 pub mod filters;
 pub mod producer;
+pub mod util;
