+++ crates/sdk/Cargo.toml
@@ -21,7 +21,6 @@ gadget-keystore = { workspace = true }
 gadget-clients = { workspace = true }
 gadget-crypto = { workspace = true }
 gadget-logging = { workspace = true }
-gadget-utils = { workspace = true }
 gadget-std = { workspace = true }
 # The keystore context is the only one that doesn't bring in extra dependencies, so we can always include it
 gadget-contexts = { workspace = true, features = ["keystore"] }
@@ -139,7 +138,6 @@ tangle = [
     "gadget-contexts/tangle",
     "gadget-context-derive?/tangle",
     "gadget-testing-utils?/tangle",
-    "gadget-utils/tangle",
     "blueprint-runner/tangle",
 ]
 
@@ -151,7 +149,6 @@ evm = [
     "gadget-clients/evm",
     "gadget-contexts/evm",
     "gadget-crypto-tangle-pair-signer?/evm",
-    "gadget-utils/evm",
     "blueprint-macros?/evm",
     "gadget-testing-utils?/anvil",
 ]
