+++ crates/sdk/src/lib.rs
@@ -22,8 +22,6 @@ pub use gadget_logging as logging;
 pub use gadget_clients as clients;
 pub use gadget_contexts as contexts;
 
-pub use gadget_utils as utils;
-
 // TODO(serial): remove
 pub use async_trait;
 pub use gadget_keystore as keystore;
