+++ crates/testing-utils/eigenlayer/src/env.rs
@@ -1,13 +1,13 @@
 use alloy_primitives::Uint;
 use alloy_primitives::{Address, address};
 use alloy_provider::Provider;
+use blueprint_evm_extra::util::get_provider_http;
 use blueprint_runner::eigenlayer::config::EigenlayerProtocolSettings;
 use eigensdk::utils::slashing::middleware::registrycoordinator::ISlashingRegistryCoordinatorTypes::OperatorSetParam;
 use eigensdk::utils::slashing::middleware::registrycoordinator::IStakeRegistryTypes::StrategyParams;
 use eigensdk::utils::slashing::middleware::registrycoordinator::RegistryCoordinator;
 use gadget_anvil_testing_utils::get_receipt;
 use gadget_logging::info;
-use gadget_utils::evm::get_provider_http;
 
 /// The default Allocation Manager address on our testnet
 pub const ALLOCATION_MANAGER_ADDR: Address = address!("8A791620dd6260079BF849Dc5567aDC3F2FDC318");
