--- crates/utils/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "gadget-utils"
-version = "0.1.0"
-description = "Utilities metapackage Tangle Blueprints"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-gadget-utils-eigenlayer = { workspace = true, optional = true }
-gadget-utils-evm = { workspace = true, optional = true }
-gadget-utils-tangle = { workspace = true, optional = true }
-
-[features]
-default = ["std", "eigenlayer", "evm", "tangle"]
-std = [
-	"gadget-utils-eigenlayer?/std",
-	"gadget-utils-evm?/std",
-	"gadget-utils-tangle?/std",
-]
-web = [
-	"gadget-utils-tangle?/web"
-]
-
-eigenlayer = ["dep:gadget-utils-eigenlayer"]
-evm = ["dep:gadget-utils-evm"]
-tangle = ["dep:gadget-utils-tangle"]
