--- crates/utils/eigenlayer/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "gadget-utils-eigenlayer"
-version = "0.1.0"
-description = "EigenLayer utilites for Tangle Blueprints"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-gadget-utils-evm = { workspace = true }
-alloy-contract = { workspace = true, default-features = false }
-alloy-network = { workspace = true, default-features = false }
-alloy-primitives = { workspace = true, default-features = false }
-alloy-provider = { workspace = true, default-features = false }
-alloy-transport = { workspace = true, default-features = false }
-eigensdk = { workspace = true, default-features = false, features = ["utils"] }
-thiserror = { workspace = true }
-
-[features]
-default = ["std"]
-std = [
-    "gadget-utils-evm/std",
-    "alloy-primitives/std",
-]
