--- crates/utils/src/lib.rs
@@ -1,8 +1,0 @@
-#[cfg(feature = "eigenlayer")]
-pub use gadget_utils_eigenlayer as eigenlayer;
-
-#[cfg(feature = "tangle")]
-pub use gadget_utils_tangle as tangle;
-
-#[cfg(feature = "evm")]
-pub use gadget_utils_evm as evm;
