--- crates/utils/tangle/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "gadget-utils-tangle"
-version = "0.1.0"
-description = "Tangle utilities for Tangle Blueprints"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-async-trait = { workspace = true, default-features = false }
-gadget-std = { workspace = true, default-features = false }
-gadget-logging = { workspace = true, default-features = false }
-tangle-subxt = { workspace = true }
-tracing = { workspace = true, default-features = false, features = ["attributes"] }
-
-[dev-dependencies]
-gadget-testing-utils = { workspace = true, features = ["tangle"] }
-tempfile = { workspace = true }
-color-eyre = { workspace = true }
-tokio = { workspace = true }
-
-[features]
-default = ["std"]
-std = [
-	"gadget-std/std",
-	"gadget-logging/std",
-	"tangle-subxt/std",
-]
-web = ["tangle-subxt/web"]
